fix: migration assumes fof/follow-tags is installed (#49)
* fix: migration assumes fof/follow-tags is installed, add very basic initial test

* remove phpunit cache

* add phpunit.result.cache

* Apply fixes from StyleCI

* enable testing

---------

+++ migrations/2023_09_12_000000_add_language_id_to_tag_user_table.php
@@ -14,12 +14,14 @@ use Illuminate\Database\Schema\Builder;
 
 return [
     'up' => function (Builder $schema) {
-        $schema->table('tag_user', function (Blueprint $table) {
-            $table->integer('dl_language_id')->unsigned()->nullable();
+        $schema->table('tag_user', function (Blueprint $table) use ($schema) {
+            if (!$schema->hasColumn('tag_user', 'dl_language_id')) {
+                $table->integer('dl_language_id')->unsigned()->nullable();
 
-            $table->foreign('dl_language_id')->references('id')->on('discussion_languages')->onDelete('set null');
+                $table->foreign('dl_language_id')->references('id')->on('discussion_languages')->onDelete('set null');
+            }
 
-            $table->index(['user_id', 'subscription', 'dl_language_id']);
+            $table->index(['user_id', 'dl_language_id']);
             $table->index(['dl_language_id']);
         });
     },
@@ -29,7 +31,7 @@ return [
             $table->dropForeign(['dl_language_id']);
 
             $table->dropIndex(['dl_language_id']);
-            $table->dropIndex(['user_id', 'subscription', 'dl_language_id']);
+            $table->dropIndex(['user_id', 'dl_language_id']);
 
             $table->dropColumn('dl_language_id');
         });
